Analysis  of Temperature Rise in PMSM
This file contains a detailed analysis of factors leading to temperature rise in Permanent Magnet Synchronous Motors (PMSM). The file aims at analyzing the factors which contribute to the demagnetization of permanent magnets present as the rotor of the machine leading to temperature rise in the machine parts resulting in decreased efficiency. This file also contains a detailed HTML report on the dataset used in the analysis of the temperature rise of PMSM.